Backported changes r3803, 3818 from stable to trunk, reviewed by David Chisnall and Niels Grewe. http://review.etoileos.com/r/35/
NOTE: The changes were wrongly committed to stable first, so I apply them back 
to trunk in the reverse way.

Modified to eliminate the use of the WorkspaceCommKit which is just overkill
for the interaction with System. Added the possibility to specify another 
session manager through a default. Written documentation in README about that.

git-svn-id: http://svn.gna.org/svn/etoile/trunk/Etoile@3819 7b3f36aa-e6db-49c1-8b4c-66eac1997e64