Cleaned and fixed the code that looks up the hit test view in -trySendEvent:toWidgetViewOfItem:.
Previously the looked up hit test view was just the outermost view in the 
decorator chain bound to the item. Hence the hit test view was not the right 
one when more than one decorator were set on a layout item, only the location 
relative to the decorator item was correctly computed.


git-svn-id: http://svn.gna.org/svn/etoile/trunk/Etoile@4765 7b3f36aa-e6db-49c1-8b4c-66eac1997e64