Fixed incorrect autoresizing of the content view in a supervisor view. The problem was caused by not updating the content frame size to match the supervisor view frame size, when the content view is inserted.
Written several tests to check the sizing of the content view.

Until now, this bug was invisible because -[ETLayoutItemGroup setLayoutView:] 
was making the adjustment by itself, and -[ETLayoutItem setView:] was updating 
newView frame by itself to match the item (and supervisor view) frame.

This issue was revealed by moving layout view insertion from 
-renderLayoutItems:isNewContent: to -[ETLayout setUp].  -setUp is called by 
-[ETLayoutItemGroup setLayout:] and at this point any scroll decorator in use 
is not yet removed, that's why the layout item frame set on the layout view in 
-setLayoutView: was then not matching the supervisor view frame but the decorator 
frame.


git-svn-id: http://svn.gna.org/svn/etoile/trunk/Etoile@4914 7b3f36aa-e6db-49c1-8b4c-66eac1997e64