Fixed two broken collection HOM tests.
In ETHOMFilterCollectionWithBlockOrInvocationAndTarget(), elementHandler IMP 
was called without casting filterResult to BOOL. As a result, filterResult as a 
long long was overwriting the next parameter, hence 'snapshot' was nil in 
-[NSDictionary(ETCollectionHOM) placeObject:atIndex:inCollection:basedOnFilter:withSnapshot:].

In ETHOMMapCollectionWithBlockOrInvocationToTarget, a map on a mutable 
collection was attempting to mutate the enumerated collection which caused the 
enumerator to raise an exception. For a map in place, the enumeration is now 
done on a temporary snapshot of the initial collection.

Note: two UTI tests are still broken.


git-svn-id: http://svn.gna.org/svn/etoile/trunk/Etoile@5016 7b3f36aa-e6db-49c1-8b4c-66eac1997e64