Finished the layout item tree copying support. You can now clone any part of the UI at runtime :-) Some complex layouts such as ETCompositeLayout are not well supported yet.
Finished -[ETLayoutItemGroup copyWithZone:] to set the parent item, update the 
layout if needed and recreate the view hierarchy in the copy.
Also fixed -itemAtPath: to use 'identifier' and not 'name' as path component.

Added -setUpCopyWithLayoutContext: to ETLayout. 
Fixed -copyWithZone: to copy the root item too.

Wrote new tests that checks the copying of layout item tree.


git-svn-id: http://svn.gna.org/svn/etoile/trunk/Etoile@5271 7b3f36aa-e6db-49c1-8b4c-66eac1997e64