Rewrote the field editor management to be much simpler and more robust (we simply insert it as a subview of the window content view now and we override the instrument in use with our own until the editing ends).
Added -hitTestFieldEditorWithEvent: to ETWindowItem and modified 
-hitTestWithEvent: and -updateActiveInstrumentWithEvent: to use it.
Rewrote -setActiveFieldEditor:editedItem: -removeActiveFieldEditorItem: to 
implement the behavior described above.

Simplified -[ETActionHandler beginEditingItem:property:inRect:] a bit and 
tweaked it to prevent crash when the property is not editable or the value is nil.

Fixed -[ETEventProcessor processMouseEvent:] to correctly update the 
activeInstrument local variable in the mouse down case.


git-svn-id: http://svn.gna.org/svn/etoile/trunk/Etoile@5866 7b3f36aa-e6db-49c1-8b4c-66eac1997e64