More tidying.
Note: Testing this code showed some bugs in clang (my fault - sorry!), which are fixed in r101759.  It is strongly recommended that you upgrade clang to this revision or newer.  There is also a bug in mainline clang preventing his from compiling.  You can work around it, but it's worth waiting for Douglas to fix it properly.



git-svn-id: http://svn.gna.org/svn/etoile/trunk/Etoile@6041 7b3f36aa-e6db-49c1-8b4c-66eac1997e64