Removed some obsolete bits of LK CodeGen.  Fixed on latest LLVM.
More refactoring and tidying.  Now using a slightly more aggressive set of optimisations.  Used the new (not yet committed - should go in soon) support in LLVM for inlining stuff from external modules.  Now MsgSendSmallInt is statically compiled and linked into LKRuntime, and the required bits are inlined where they are needed.

Note: Stripping debug info is only required in the JIT until LLVM PR6981 is fixed (this was the cause of the earlier crashes).



git-svn-id: http://svn.gna.org/svn/etoile/trunk/Etoile@6098 7b3f36aa-e6db-49c1-8b4c-66eac1997e64