Make ETException work with the new ABI.
Quentin: Do we actually want to enable this?  I wrote it more as an exercise to see if it would work than as a sensible thing.  I think I'd rather we not enable it in this version, and rewrite it for the next release to use an explicit resume / restart mechanism that is only invoked when raising ETException instances, not try to overload NSException with all of this stuff.



git-svn-id: http://svn.gna.org/svn/etoile/trunk/Etoile@6991 7b3f36aa-e6db-49c1-8b4c-66eac1997e64