Make EtoileFoundation able to be compiled in ARC mode (if you run the rewriter, which I haven't done).
Removed two files from the build:

NSObject+Prototypes.m: This reimplements a load of stuff that the runtime implements.  It's ugly.  We should move 90% of it into the runtime.  

ETThread.m: As far as I can tell, this is a partial reimplementation of NSThread, with a different name.  No idea why I wrote it - I can only assume I was drunk at the time.



git-svn-id: http://svn.gna.org/svn/etoile/trunk/Etoile@7038 7b3f36aa-e6db-49c1-8b4c-66eac1997e64