Put the new db schema in use.
Massively reworked how the history is stored and accessed, it now revolves around commit 
tracks and persistent roots. Root objects are pretty much reduced to a small implementation 
detail, and don't play a significant role in the history model.

We can now implement branching and cheap copies at a high-level in COCommitTrack and 
related API.

Note: Some serious cleaning is still needed in COStore API and how various classes interact 
with it.


git-svn-id: http://svn.gna.org/svn/etoile/trunk/Etoile@7937 7b3f36aa-e6db-49c1-8b4c-66eac1997e64